turning encoder updates
The code is setup for using a secondary turning encoder wired into the turning sparkmax, e.g. if the CANcoder were connected to the sparkmax, this would work. The code WILL NOT work with the built-in encoder on the turning neo. It wraps around and cannot be used in absolute mode like the code requires. It might be possible to stop configuring it in absolute mode and keep track of revolutions of the turning motor and translate it to an absolute position using the turning gear ratio (similar to the drive ratio giving the pid velocity in meters).