restore swerve module first attempt
Restore the swerve module code used in our first attempt to get swerve working. Somehow we ended up with two copies of the MaxSwerve module code.

After experimenting with the Rev MaxSwerve template code, it is easier to see what code can be reused and where the bugs are in the first attempt. It should be possible to replace the MaxSwerve m_turningPIDController (an analog encoder connected to the encoder port on the SparkMax) with a PIDController java object using the CANCoder as input. It will probably be necessary to add a feedforward factor for more responsive turning.